functional python script
